+++ externs.js
@@ -1,3 +1,3 @@
 const createSubConfig = require('./utils/create-sub-config');
 
-module.exports = createSubConfig('zombiebox');
+module.exports = createSubConfig('externs');