+++ overrides/externs.js
@@ -1,0 +1,10 @@
+module.exports = {
+	rules: {
+		'prefer-const': 'off', // "const" is forbidden in externs
+		'no-unused-vars': 'off',
+		'no-empty-function': 'off',
+		'interfaced/valid-jsdoc': 'off',
+		'interfaced/no-unused-expressions': 'off',
+		'interfaced/no-empty-method': 'off'
+	}
+};